remove core knbos
