--- packages/widgetbook_core/lib/src/settings/features/features.dart
@@ -1,2 +1,0 @@
-export './knobs/knobs.dart';
-export './properties/properties.dart';
