--- packages/widgetbook_core/lib/src/settings/features/knobs/knob_padding.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class KnobPadding extends StatelessWidget {
-  const KnobPadding({
-    required this.child,
-    super.key,
-  });
-
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 8),
-      child: child,
-    );
-  }
-}
