--- packages/widgetbook_core/lib/src/settings/features/knobs/knobs.dart
@@ -1,7 +1,0 @@
-export './bool_knob.dart';
-export './color_knob.dart';
-export './number_knob.dart';
-export './option_knob.dart';
-export './slider_knob.dart';
-export './text_knob.dart';
-export './knob_property.dart';
