--- packages/widgetbook_core/lib/src/settings/features/properties/properties.dart
@@ -1,1 +1,0 @@
-
