+++ packages/widgetbook_core/lib/src/settings/settings.dart
@@ -1,3 +1,2 @@
-export './features/features.dart';
 export './models/models.dart';
 export './widgets/widgets.dart';
