+++ packages/widgetbook_core/lib/src/settings/widgets/widgets.dart
@@ -1,5 +1,6 @@
 export './complex_setting.dart';
 export './dropdown_setting.dart';
+export './knob_property.dart';
 export './setting.dart';
 export './setting_group.dart';
 export './setting_header.dart';
