--- widgetbook_for_widgetbook/lib/settings/features/knobs/bool_knob.widgetbook.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:widgetbook_annotation/widgetbook_annotation.dart';
-import 'package:widgetbook_core/widgetbook_core.dart';
-
-@UseCase(name: 'Default', type: BoolKnob)
-Widget boolKnob(BuildContext context) {
-  return const BoolKnob(
-    name: 'Is enabled',
-    description: 'This is a description.',
-    value: true,
-  );
-}
-
-@UseCase(name: 'Default', type: NullableBoolKnob)
-Widget nullableBoolKnob(BuildContext context) {
-  return const NullableBoolKnob(
-    name: 'Is enabled',
-    description: 'This is a description.',
-    value: true,
-  );
-}
