+++ widgetbook_for_widgetbook/lib/settings/widgets/settings_panel.widgetbook.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:widgetbook_annotation/widgetbook_annotation.dart';
+import 'package:widgetbook_core/widgetbook_core.dart';
+
+@UseCase(name: 'Default', type: SettingsPanel)
+Widget settingsPanel(BuildContext context) {
+  return SettingsPanel(
+    settings: [
+      SettingsPanelData(
+        name: 'Properties',
+        settings: [],
+      ),
+      SettingsPanelData(
+        name: 'Knobs',
+        settings: [],
+      ),
+    ],
+  );
+}
